pyproject.toml
[project]
name = "sniplink"
version = "0.1.0"
description = "This is a solution to the URL shortener project on roadmap.sh : https://roadmap.sh/projects/url-shortening-service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest-django>=4.11.1",
    "ruff>=0.13.1",
    "django-stubs>=4.2.7",
    "djangorestframework-stubs",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"
